changed ParsedInputs to InputService, refactored test cases
