+++ src/service/arguments/cli/tools/tool-subcommand-factory.ts
@@ -2,7 +2,7 @@ import { Command } from "commander";
 import { CommandConstructor } from "@bc/service/arguments/cli/command-constructor";
 import { CLIActionType, ToolType } from "@bc/domain/cli";
 import { ProjectListCommand } from "@bc/service/arguments/cli/tools/project-list";
-import { ParsedInputs } from "@bc/service/inputs/parsed-inputs";
+import { InputService } from "@bc/service/inputs/input-service";
 import Container from "typedi";
 
 
@@ -31,7 +31,7 @@ export class ToolSubCommandFactory {
             .option("-t, --token <token>", "The GITHUB_TOKEN. It can be set as an environment variable instead")
             .option("-d, --debug", "Set debugging mode to true", false)
             .action((options) => {
-                const parsedInputs = Container.get(ParsedInputs);
+                const parsedInputs = Container.get(InputService);
                 parsedInputs.updateInputs({...options, CLICommand: CLIActionType.TOOLS, CLISubCommand: toolType});
             });
